+++ app/Http/Controllers/Api/V1/TicketController.php
@@ -8,10 +8,8 @@ use App\Http\Requests\Api\V1\UpdateTicketRequest;
 use App\Http\Requests\Api\V1\ReplaceTicketRequest;
 use App\Http\Resources\V1\TicketResource;
 use App\Models\Ticket;
-use App\Models\User;
 use App\Policies\V1\TicketPolicy;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
-use Illuminate\Auth\Access\AuthorizationException;
 
 class TicketController extends ApiController
 {
@@ -30,15 +28,11 @@ class TicketController extends ApiController
 	 */
 	public function store(StoreTicketRequest $request)
 	{
-		try {
-			//policy
-			$this->isAble('store', Ticket::class);
-
+		if ($this->isAble('store', Ticket::class)) {
 			return new TicketResource(Ticket::create($request->mappedAttributes()));
-
-		} catch (AuthorizationException $exception) {
-			return $this->error('You are not authorized to create this resource', 403);
 		}
+
+		return $this->error('You are not authorized to create this resource', 403);
 	}
 
 	/**
@@ -69,15 +63,14 @@ class TicketController extends ApiController
 			$ticket = Ticket::findOrFail($ticketId);
 
 			//policy
-			$this->isAble('update', $ticket);
+			if ($this->isAble('update', $ticket)) {
+				$ticket->update($request->mappedAttributes());
 
-			$ticket->update($request->mappedAttributes());
-
-			return new TicketResource($ticket);
+				return new TicketResource($ticket);
+			}
+			return $this->error('You are not authorized to update this resource', 403);
 		} catch (ModelNotFoundException $exception) {
 			return $this->error('Ticket cannot be found', 404);
-		} catch (AuthorizationException $exception) {
-			return $this->error('You are not authorized to update this resource', 403);
 		}
 	}
 
@@ -91,16 +84,17 @@ class TicketController extends ApiController
 			$ticket = Ticket::findOrFail($ticketId);
 
 			//policy
-			$this->isAble('replace', $ticket);
+			if ($this->isAble('replace', $ticket)) {
+				$ticket->update($request->mappedAttributes());
 
-			$ticket->update($request->mappedAttributes());
+				return new TicketResource($ticket);
+			}
 
-			return new TicketResource($ticket);
+			return $this->error('You are not authorized to update this resource', 403);
 		} catch (ModelNotFoundException $exception) {
 			return $this->error('Ticket cannot be found', 404);
 		}
 	}
-
 
 	/**
 	 * Remove the specified resource from storage.
@@ -111,11 +105,13 @@ class TicketController extends ApiController
 			$ticket = Ticket::findOrFail($ticketId);
 
 			//policy
-			$this->isAble('delete', $ticket);
+			if ($this->isAble('delete', $ticket)) {
+				$ticket->delete();
 
-			$ticket->delete();
+				return $this->ok('Ticket successfully deleted');
+			}
 
-			return $this->ok('Ticket successfully deleted');
+			return $this->error('You are not authorized to delete this resource', 403);
 		} catch (ModelNotFoundException $exception) {
 			return $this->error('Ticket cannot be found', 404);
 		}
